Splits the TempActionService.cs singleton so that TempActionChecker.cs now has the actual timed loop check which the service calls through a better timer
